MAJOR REWRITE
A major rewrite. I had been fixing and rewriting things on this particular script for awhile, but I hadn't published any because I didn't want to go through the hassle of updating the instructions yet. Mainly because it now requires Subtitle Edit and FastText, both of which, can be a big hassle to setup depending on the OS.