Added forced track status, lessened amount of discord messages to only errors and changes, logic for checking remaining subs in mkv before external ones
So currently, I have logic that, in the event that detection falls below the required percentage, it will do two things. 

First, it will check the rest of the remaining subtitle tracks within the internal mkv file, parse them into a subtitle array, check for duplicates, then retest. This will eliminate the romaji that throws off the language detection, then recheck it without the duplicates. 

Second, if it still fails using the remaining tracks within the file, it will parse the release group of the file (ASSUMING THE RELEASE NAME IS ON THE END OF THE FILE, FORMATTED AS "filename-releasegroup.mkv"). Find additional files that are also by that releaser, and use those to find the duplicates.

The reason I wrote this logic was because I found that when the Signs & Songs track had romaji for both the intro and outro of an anime episode, it caused detection to fall behind my set threshold of 70%. So by finding duplicates between the Signs & Songs track and another track, it can eliminate the romaji, and then retest it.

I have currently commented this logic out as I still need to test it more to make sure everything is perfect before having it run by default. There's currently a couple challenges I need to tackle with it. 

That and I'd like to eventually also do a huge code rewrite of everything, cleaner, simpler, and using classes.